Fixed types for error for repositories/pullrequests/entities ducks
